initial work

+++ gpu/gpu_search_v2.cu
@@ -98,12 +98,12 @@ __global__ void gpuSearchBtree(unsigned char * btree_trie_mem, unsigned int * fi
         __syncthreads();
 
         //Set the start index
-        unsigned int current_btree_start = *next_level*4;
+        uint64_t current_btree_start = *next_level*4;
         current_ngram++;
         key = keys_shared[current_ngram];
 
         //Some necessary variables
-        unsigned int updated_idx;
+        uint64_t updated_idx;
         unsigned int size;
 
         //Current_btree_start == 0 means that we had UNK key (vocabID 1 which wasn't found, so we should directly go to backoff
